Avoid re-render of the whole app when the DayTimeContext changes.
