+++ tv/pages/_app.tsx
@@ -1,43 +1,24 @@
-import { useState } from 'react';
 import '../styles/globals.css';
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
 import Navigation from '../components/Navigation';
-import {
-  DayTimeContext,
-  DayTimeContextValue,
-  dayTimeDefaultValue,
-} from '../contexts/DayTimeContext';
+import { DayTimeContextProvider } from '../contexts/DayTimeContext';
 
-function App({ Component, pageProps }: AppProps) {
-  const [dayTimeState, setDayTimeState] = useState(dayTimeDefaultValue);
+const App = ({ Component, pageProps }: AppProps) => (
+  <div className="screen">
+    <Head>
+      <title>PLAY! Summit TV</title>
+      <meta name="description" content="PLAY! Summit TV displays" />
+      <link rel="icon" href="/favicon.ico" />
+    </Head>
 
-  // TODO: Will probably re-render due to using a custom re-created object.
-  const dayTimeContextValue: DayTimeContextValue = {
-    dayTime: dayTimeState,
-    setDayTime: (day: string, time: string) => {
-      setDayTimeState({
-        day,
-        time,
-      });
-    },
-  };
+    <main>
+      <DayTimeContextProvider>
+        <Navigation />
+        <Component {...pageProps} />
+      </DayTimeContextProvider>
+    </main>
+  </div>
+);
 
-  return (
-    <div className="screen">
-      <Head>
-        <title>PLAY! Summit TV</title>
-        <meta name="description" content="PLAY! Summit TV displays" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main>
-        <DayTimeContext.Provider value={dayTimeContextValue}>
-          <Navigation />
-          <Component {...pageProps} />
-        </DayTimeContext.Provider>
-      </main>
-    </div>
-  );
-}
 export default App;
